backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.4.0
python-dotenv==1.0.0
openai==1.54.3
httpx==0.27.0
PyJWT==2.8.0
requests==2.31.0
google-generativeai==0.8.3